Test/Test/requirements.txt
streamlit
boto3
test
langchain.llms.bedrock
langchain.memory
langchain.chains